+++ libraries/I2C_LCD/I2C_LCD_special_chars.h
@@ -8,13 +8,17 @@
 //     URL: https://github.com/RobTillaart/I2C_LCD
 
 
-//  SPECIAL CHARS,  might not work on all displays.
+//  SPECIAL CHARS,
+//  will only work on displays with ROM CODE A00
+//
 
 const char LCD_ALPHA    = 0xE0;
 const char LCD_BETA     = 0xE2;
 const char LCD_EPSILON  = 0xE3;
 const char LCD_MU       = 0xE4;
 const char LCD_RHO      = 0xE5;
+const char LCD_SQROOT   = 0xE7;
+
 const char LCD_THETA    = 0xF2;
 const char LCD_INFINITY = 0xF3;
 const char LCD_OHM      = 0xF4;