+++ libraries/I2C_LCD/CHANGELOG.md
@@ -6,15 +6,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.4] - 2023-12-28
+- changed return type of **begin()**. Returns false if LCD not found on I2C bus.
+- made initialization delay(100) in **begin()** a bit smarter.
+- add **repeat(c, times)**
+- fix **center()** for non 20x4 LCD.
+- add **I2C_LCD_custom_chars.h** with examples.
+- add examples
+
+
 ## [0.1.3] - 2023-12-22
-- add support for **16x4** + **10x4** display
+- add support for **16x4** and **10x4** display.
 - minimized footprint **setCursor()**
 - add parameter to **moveCursorLeft(uint8_t n = 1)**
 - add parameter to **moveCursorRight(uint8_t n = 1)**
 - improve timing
 - update examples
 - minor edits
-
 
 ## [0.1.2] - 2023-12-21
 - implement polarity backlight (and removed policy)